AMG runs in parallel now.
There was huge bug in the numbering of the aggregates:

Due to the coarsening of the index set the matrix on the coarse level are ordered like this. First come the row corresponding to vertices owned by the process
(0, 1, ..., n) and the the rows owned by other process (n+1, n+2, ..., N).

Unfortunately during the setup of the sparsity pattern of the matrix this fact was neglected :-(

Fixed this now.

[[Imported from SVN: r500]]