Add a mode to SuperLU where apply uses freshly allocated vectors.
Previously the two vectors are allocated in
the first call to apply. These got resused in subsequent calls to apply
and are deallocated in the destructor. Using the new mode these vectors are allocated
at the beginning and deallocated at the end of each apply method. This allows
using the same instance of superlu from different threads.

Patch provided by Arne Morten Kvarving. I just added some more
documemtation and a missing initialization.

[[Imported from SVN: r1754]]