[Merge][Bugfix] r1895 from branches/release-cmake-2.2:  Prevent definition of HAVE_BOOST without value due to race conditions in some autoconf versions.
For some autoconf versions there was a race condition between ax_boost_base.m4 and dune_boost_base.m4 resulting in HAVE_BOOST being defined with no value (instead of ENABLE_BOOST). For others there were even two definitions of it (one with and one without value). 

This patch renames the symbol being defined in dune_boost_base.m4 to HAVE_DUNE_BOOST and defines a value for it if boost is found.

Please change your code such the preprocessor directives checking for boost from
#if HAVE_BOOST
to
#if HAVE_DUNE_BOOST

Resolves flyspray 1265 and presumably 1249.

This time the patch is clean (as opposed to r1896 which was reverted in r1901).

[[Imported from SVN: r1902]]