[Bugfix][SuperLU] Use a pointer and not a reference for pointing.
A reference is an lvalue. Assigning to it later changes the state
of the rhe referencee (and not the reference). Therefore the code
actually stored references to the right hand and left hand side.
In the case where we did not want to reuse vectors we created two
dense SuperMatrices to used in this apply call and then assigned
them to the references. In fact we actually overwrote the right
and left hand side by doing this.

This patch uses pointers to accompish the intended behaviour, i.e.
point to the correct dense matrices, the ones stored in SuperLU
or the ones stored in SuperLU::apply.