[Bugfix] Prevent implicite conversion Matrix->MatrixAdapter.
MatrixAdapter uses a const reference to store a reference
to the matrix it is working on. Without this patch it was
possible to pass the matrix to constructor actually taking
a MatrixAdapter as its argument. This resulted in a const
reference to a temporary MatrixAdapter that died after the
constructor call.

This patch request explicitly calling the constructor and
thus fixes this issue.