[Vector] WIP version of host vector with TBB and kernel support
This patch is mostly just a snapshot of the host vector implementation.
It should mostly work, provided you have all the dependencies (I've tested
it with GCC 4.8 and TBB 4.1_4).

The implementation uses TBB to multithread algebraic operations. The actual
operation kernels are implemented in the new Kernel library in dune-common,
which has several advantages:

- I noticed that inlining of the kernels messes up the auto-vectorizers of both
  GCC and ICC, so the kernels have to be in external functions anyway.

- Having the kernels in dune-common allows a lot more reuse.

- Precompiling the kernels can save on compilation time (and the mechanism in
  dune-common also allows for specializing specific variants of the kernels,
  if someone feels inclined to play around with SIMD intrinsics).

I've tried designing in support for mixed precision, but that is untested as of
now.

You need to use one of the special allocators from dune/common/kernel/allocator.hh
because those export some required extra information (alignment and block size).