[Containers] Move problematic forward declarations with default template arguments to separate header
C++ requires default template arguments to appear as part of the first
declaration of a template, which creates a problem in the interactions
between the new containers (which need a default argument for their
memory domain) and PDELab (which has to be able to work with those
containers as incomplete types).

This patch resolves the problem by moving the initial
declarations (including the default arguments) to a separate header file.