[twolevelmethod][bugfix] Copy coarse level solver for thread safety. (2nd tparam of TwolevelMethod changed!)
Previously all copies of the twolevel method used the same
coarse level solver. In the case of AMG this resulted in wrong
results as it stores internal data across several applications.
This patch now always copies the internal coarse level data.

In addition the left hand side given to AMG as a coarse level solver
is copied to use it for the post method in the constructor.

Note that the meaning of the template parameters of TwolevelMethod is
changed. Its second parameter is now the type of the coarse level
solver rather than the coarse level linear operator. The latter is
deduced from the previous now. This was needed to supportt copying of
the coarse level solver.