[Testsystem] Change test systems to allow easy use of different allocators
The current ISTL test system relies heavily on test problems generated
with setupAnisotropic2d() from dune/istl/paamg/anisotropic.hh . The code
there was written in a fashion that forbid the use of an allocator different
than std::allocator on the BcrsMatrix.

This patch fixes this issue by giving the matrix type as template parameter
to the matrix setup instead of blocksize and fieldtype. This also makes
code easier to read. All (hopefully) tests in ISTL are adpated to the change.