[bugfix] Fixes usage of scotch (parmetis/metis)
There were problems with repartition.hh when using the original parmetis. The cause was
that for scotch we need to require both parmetis.h and metis.h, which broke backwards
compatibility for the original parmetis (which worked without metis, too). With the patch
we check whether parmetis.h is from scotch and if that is the case also require metis. If not,
metis suffices.

(cherry picked from commit ba0f24fae420a74d778f4ddadf54e3cb02711129)
Signed-off-by: Markus Blatt <markus@dr-blatt.de>