[bugfix,parmetis] Always typedef idxtype.
The type idxtype is used by header repartition even if parmetis is not
used. Therefore it has to be typedefed meaningful in this case,
too. This was not the case before and is accomplished by this commit.

(cherry picked from commit f8a19fb0b9e03a719fa21c21177a878cc5d0fb7a)
Signed-off-by: Markus Blatt <markus@dr-blatt.de>