[BCRSMatrix] Drop over-eager check in row access operator[]
When DUNE_ISTL_WITH_CHECKING is enabled, the BCRSMatrix throws an
exception when trying to access a row that has no columns. This creates
a problem in PDELab when using nested BCRSMatrices (as some people do
for operator splitting in PDELab), where an off-diagonal BCRSMatrix
block might be completely empty (i.e. have a correctly set up structure,
but every row in the off-diagonal BCRSMatrix is empty. If the PDELab
assembler then attempts to set a Dirichlet constraint, it has to access
the row to set all columns in that row to zero, but the check in
operator[] prevents it from doing so.

So this patch just drops that check.