Choose different codepath if METIS 5 is found.
Updated how METIS_PartGraphRecursive is called in METIS version 5.
In the new parameters NULL value is used where possible. Only one
new parameter could not be set to NULL, it is the "ncon" (second argument).
In the manual it says: "ncon: The number of balancing constraints.
It should be at least 1." so 1 is used as its value.

Also options with the new layout are passed in METIS_PartGraphRecursive
initialized with the default values(default values also used in the
current code: options[0]=0). Only one option was added manually:
the METIS_OPTION_NUMBERING = numflag which is the same numflag parameter
passed in the old version of METIS_PartGraphRecursive function.

Signed-off-by: Christoph Grüninger <gruenich@dune-project.org>