[bugfix] "properly" from field_type to int
we have to retrieve an int which is stored in field_type. If the field_type is std::complex, we
have to access the real-component, if it is double/float, we don't have to do anything
special. We use std::real(.) to get the correct real value. This function (with its appropriate
overloads) is available from c++11 on, but this part was already supported by the ancient
g++ version we supprt ;-)