[release] Do not assume a symmetric sparsity pattern in SymmetricDependency.
Previously, we assumed that if a_{ij} is stored in the sparse matrix
then a_{ji} must be stored also and used mat[i][j] to access. If the entry
was not stored then an exception somewhere in basearray.hh was thrown.
Unfortunately knowing the cause for this exception is insider knowledge,
seldomly leaked ;). In addition there is now reason to throw an expection here.

With this commit we the find method on the row to search for the entry a_{ji}.
If it is not present we simply treat it as being zero.

(cherry picked from commit 05c0a63a4872eebdb8ee5223cb4c80c35b4060dd)
Signed-off-by: Carsten Gräser <graeser@dune-project.org>