Makes vectorcommtest independent of dune-istl data types.
This test actually does test the parallel functionality
that is now contained in dune-common for quite some time.
While it is using BlockVector this is not needed by the test.
It works with any container supporting operator[].

Therefore this commit gets rid of BlockVector and uses std::vector
instead. This will allow us to move the test to dune-core subsequently.