Add unit tests for the MultiTypeBlockVector and MultiTypeBlockMatrix classes
Both tests mainly test whether all interface methods compile, and whether
the resulting code does not crash (it doesn't).  A large part of the matrix
test has been copied from the file dune/istl/tutorial/example.cc, which
is really much more of a test than an example.

That test for MultiTypeBlockMatrix seems to trigger a valgrind complaint
inside BCRSMatrix.  I'll need to investigate this.

Both test executables are built even if boost fusion is not found.  In that
case, both simply return '77'.  I am aware that this may not be the official
way to handle the boost dependency, but I didn't want to spend to much time
on the corresponding cmake way to skip the test, given that my plan is
to remove the boost dependency anyway.