[istl][multitype] fix solvers for MultiType*
the old code didn't work (at least not with clang). The reason is that constructs like
template<typename... Arg1, typename... Arg2> don't work, as the  compiler doesn't know
where to split the list of arguments. The whole algo_itsteps got reworked to allow for an
easier specialization for MultiTypeBlockMatrix. Now the struct is also parametrized by the
matrix type. By this we can easily add a specialization.