[cmake] link to arpack++ if found
The arpack++ package from Debian comes with a shared library that we
have to link to. Otherwise the linker will complain about missing
symbols when we try to build arpackpptest.
As there seem to be other versions of arpack++ that are header-only, we
only link to the shared library if it was found.