Introduce dune_add_test in dune-istl
There are some tests, which are hard to port in the paamg directory.
They want to explicitly test _without_ direct solver. Previously,
you simply did not add the flags for the direct solver packages.
add_dune_test relies on all flags, so we need a way to disable components
again. That will be implemented as part of dune_add_executable. For
now, we build the targets ourselves and hand them to dune_add_test.