Introduce file vectortest.hh, with tests for the dune-istl vector interface
So far, each dune-istl vector class had its own unit test.  They all tested, among
other things, for compliance with the dune-istl vector interface.  This lead, or had
the risk to lead, to subtle differences between the individual tests.  Also, since
there is no formal specification of the interface, it was somewhat unclear what the
interface actually requires.

This patch tries to improve the situation by providing a generic set of tests that
should be passed by anything claiming to be a dune-istl vector.  This will save work,
because it allows to centralize tests that are now distributed across the different
unit tests for the different dune-istl vectors.  Also, it is my hope that these
centralized tests can serve as some form of formalized specification of the dune-istl
vector interface.