Make class 'Matrix' work again, by not using VariableBlockVector anymore
The recent changes to VariableBlockVector broke the Matrix class -- shame on me
for testing thoroughly enough before pushing.  However, using VariableBlockVector
internally in Matrix was a hack anyway.  So, as a first step towards a cleaner
solution, this patch copies the (old) VariableBlockVector class into the file,
renames VariableBlockVector to DenseMatrixBase, and does some obvious cleanup.
This is still not particularly pretty, but better than before.