[bugfix] Cater for the case that we load balance to more processes.
One of the use cases for load balancing a matrix is reading it globally
on one process and then calling the load balancer to distribute it among
the processes. Unfortunately this did not work due to a bug that prevented
the correct remapping of the domain numbers from the load balancers. This
lead to negative rank numbers.

With this commit we use yet unassigned process numbers for domains that
are not yet mapped (due to having no dofs before the load balancing).