Remove the 'block' array
This array was a left-over from the VariableBlockVector class.  In short,
it stores the length of each matrix row.  Of course that is redundant,
because all rows in a dense matrix have the same length.

This patch makes the class allocate less memory, and use the known row
size to compute the offsets into the data array.  In principle, the code
becomes faster by this, but I am not convinced that the effect will be
measurable.

Still, even without measurable speedup: this patch is an improvement,
because it makes the code simpler.