[BCRSMatrix] Provide access to raw column and value data
This patch adds methods columnIndices() and values() that return raw
pointers to the storage of the corresponding data.

These methods are mainly intended for situations where matrices are
constructed in a non-standard way with prior knowledge, i.e. when
repartitioning parallel matrices. In this case, data is streamed into
the matrices by means of MPI calls that want to directly write to the
matrix.