[AMG] Add RepartitioningStrategy
This patch adds a repartitioning strategy class to the AMG
implementation. This class will be called during the creation of the AMG
hierarchy in a parallel setting to control the repartitioning of the
parallel program onto fewer processes. It returns whether a
repartitioning should be done and if yes, how many processes should be
part of the new partitioning.

The strategy is implemented as an abstract base class, which makes it
possible to choose a strategy at run time.

There are currently three implementations:

- DefaultRepartitioningStrategy mimicks the current behavior of the AMG
  coarsening process by always trying to assign (coarsenTarget *
  minAggregateSize) degrees of freedom to each processor. This will
  normally cause a repartitioning at each level as soon as the matrix
  has become sufficiently small.

- FixedRepartitioningStrategy simply divides the number of processes by
  "aggregationFactor" every time ((level + repartitioningShift) %
  repartitioningInterval) == 0. This gives very precise control over the
  behavior of the repartitioning, but might lead to less than optimal
  partition sizes in case of an uneven coarsening rate.

- BandedRepartitioningStrategy tries to keep the number of DOFs per
  process between minUnknownsPerProcess and maxUnknownsPerProcess.

The first strategy takes its parameters from the AMG criterion, while
the latter two each take a ParameterTree from which they will parse
their parameters.