[AMG][bugfix] Fix loop that calls pre() for all smoothers in the hierarchy
The semantics behind this code are horribly twisted (mainly due to the
fact that the hierarchy doesn't follow standard STL container
semantics), and as a result the algorithm would in certain cases try to
call pre() on a non-existent smoother, causing either a crash (still ok)
or a hang in some MPI routing (very bad...).

The new version isn't much better, but it is at least correct in all cases.