[Communication] Add commgraph-based vector redistribution
This commit adds an alternative implementation of
`RedistributeInformation` that exploits the additional structure
afforded by doing a commgraph-based repartitioning to more efficiently
repartition distributed vectors.

It uses a dedicated communicator for each partition of the vector that
gets merged onto a single rank and only requires a single
`MPI_Gatherv()` for each merge operation and one `MPI_Broadcast()` per
scatter operation.

The class takes ownership of the MPI communicator and frees it on
destruction.