[Communication] Add new implementation of commgraph-based marix repartitioning
This version differs from the existing implementation in that it fully
exploits the properties of a commgraph-based repartitioning, in
particular the fact that all rows of a given rank will move to a single
target rank. It also creates separate MPI communicators for each merged
matrix repartition, overlaps computation and communication and rebuilds
global index information locally without global communication.