[AMG] Rework repartitioning of AMG during setup
This patch replaces the repartitioning algorithm used to redistribute
the matrix to fewer ranks during the setup phase with a more efficient,
completely commgraph-based version and also makes it possible to control
the repartitioning process from the outside by passing in an object that
inherits from `RepartitioningStrategy`.

At the same time, it also simply stores the number of unknowns in a
`std::size_t` instead of a special type, as 10^19 unknowns are probably
sufficient for the near future.

The old repartitioning code is still there; in order to use it, define
the macro AMG_USE_OLD_REPARTITIONING. Note that it is not possible to
switch between the algorithms at run time as they use very different
data structures to store the information required to gather and
redistribute vectors during the actual AMG application.