[AMG][Bugfix] Fix AMG when redoing repartitioning on coarsest level
The AMG code had a bug that caused a deadlock in parallel code when it
became necessary to redo the partitioning at the coarsest level (to get
all data onto a single rank). In that case, all of the ranks that were
part of the redistributed matrix after the first repartitioning still
believed that they had the coarse solver and tried to run it.