Ignore boost::fusion 1.61 and above.
Version 1.61, unlike 1.60 and earlier versions, makes compilation
of the example fail with an error in multitypeblockvector.hh. See also
issue #18.

Since the code in question is just an example and boost::fusion is no
longer used to implement multitypeblockvectors in more recent version of
DUNE, no attempt is made to provide a proper fix.