Backward compatiblity for remaining solvers
All constructors were tested in the following sense:

They were all called in the old way but I didn't test if the new way
still works. So old constructors work but the new way could be problem
if I made a mistake in the enable if trick.

Note: I didn't adapt the deprecated constructors for amg and the
constructor where you want to build your own hierarchy.