[comment] Fix comment: This is a downcast
This is in fact an _unsafe_ downcast to a derived
class. Accessing members via this pointer is undefined
if it is not guaranteed that the passed pointer
points to an object of the _derived_ class - even
if the memory layout is the same. In this case
this static_cast acts like a reinterpret_cast.