[compatibility] allow to disable the compatibility layer
depending on a preprocessor define we switch the base classes to be
pure virtual, or implement a dummy category(). If the classes are pure
virtual the old interface with an enum is not possible anymore.

To enable/disable the backwards compatibility check we provide a cmake
flag DUNE_ISTL_SUPPORT_OLD_CATEGORY_INTERFACE, which defaults to 1
(e.g. backwards compatibility enabled).