[bugfix] Make sure to use tolower from c
Matrixmarket so far uses std::transform(..,..,.., tolower).
However it only works with the C function tolower, not
with std::tolower (which is a template). The patch explicitly
chooses the C version. (Of course, this only breaks if somebody
draws std::tolower or similar into the scope, which I  would generally discourage.
But the resulting failure can be avoided, so we should avoid it.)