Comment out warnings
These warnings are very irritating. They do not
warn about deprecated or misused feature. They are
issued any time you compile the tests. They don't
provide any useful information unless you look at the
code and have insight in the implementation detains.

To sum up: They don't look like a warning addressed to
the user, but like a reminder on an intended future
implementation addressed to the implementor. Hence
they should not be exposed to the user.

If anyone feels that this is an important user warning
feel free to re-enable it after adding enough information
making the warning actually useful.