Fix segmentation fault in AMG if matrix of one rank is empty.
For a matrix with any rows in a parallel run, The subgraph
containing only vertices associated with matrix rows owned by
the process would have 1 vertex although the actual matrix graph
would have none. This is fixed by checking the number of vertices
of the outer graph and act correctly for zero vertices.
For this we need to mark a matrix without rows built in row_wise mode
as built once the CreateIterator is constructed and correct the number
of vertices for a Subgraph of a MatrixGraph with 0 vertices.