Remove the file 'example.cc'
This file, while being called 'example', was actually a set of unit
tests.  The stuff it tested was all covered by the standard unit
tests.  Having these additional tests in a file with a misleading
name, and at a location outside of the standard test locations,
was confusing (for me, at least).  This patch simply removes
the file.

With the 'example.cc' file gone, the directory 'dune/istl/tutorial'
is empty.  Therefore this patch removes that directory, too.