Make the CreateIterator of VariableBlockVector a forward iterator
The CreateIterator of VariableBlockVector is currently an output iterator. This creates problems
with the tests, which use std::fill() with those iterators, and that function actually requires
forward iterators (because output iterators are not comparable). The libc++ implementation of clang
7 actually enforces this, causing test failures.

Looking at the implementation, I see no reason for not making the CreateIterator a forward iterator,
so this patch does just that.