Move code into the generic template class SuperLU<M>
Previously, all code was in the specialization of that class
for BCRSMatrix<FieldMatrix>.  That was overly restrictive:
most of the code does not actually rely on the precise type
of the matrix.  As a first step towards allowing more general
matrix type we move all the code into the general class.