[amg] don't copy temporary data durint copy-construction
if on the initial AMG pre was called, it is currently it not necessary to
call pre on a copy of that AMG. This is a feature that is not required
by the Preconditioner interface and thus it nothing anybody can rely on.
We now don't copy the temporary data anymore.