copyWindowStructure should not set the matrix build mode to 'row_wise'
Rather, it should take the value from the matrix that is being
copied from.

In particular, BCRSMatrix::copyWindowStructure is being called as
part of BCRSMatrix::operator=.  Previously, that method set the
matrix build mode to 'row_wise' unconditionally.  This does not
pose problems until the matrix is asked to be resized later,
because that only works for matrices with the 'random' build mode.