Cholmod: Change template specialization to vector type and extent the handling of ignored entries
We interpret Cholmod as an InverseOperator between BlockVectors. Therefore
we dump the BCRSMatrix template.

To clarify the usage of ignore nodes the apply() method now ignores the corresponding entries directly.
The information is stored in a mapping from all indices to the not ignored indices.

Also, the unit tests for cholmod are extented.