overlappingschwarztest: remove broken runtime check
This test should not be necessary anymore, since commit
05f80ffad46f5e0d5566d0a6f49589ba3be61e18 fixed the CMAKE_GUARD argument
for this test, so that it is not compiled anymore if either UMFPACK or
SuperLU are missing.

That line had to be removed as it did not compile if either UMFPACK or
SuperLU were missing. For example while I had UMFPACK installed but not
SuperLU, gcc gave me the following error message for the removed line:

…/dune/istl/test/overlappingschwarztest.cc:260:17: error: ‘HAVE_SUPERLU’ was not declared in this scope